requirements.txt
streamlit
Pillow
numpy
scikit-image
plotly
pyfiglet
